Move 'Internal.Cryptography.Pal.CertificateData' into its own file.
This moves all the managed pieces of Internal/Cryptography/Pal.OSX/CertificatePal.cs
which are not platform-specific into a separate file to allow them to be used in Mono.
